Ignore provider update diffs
Upgrade tests started failing by over-eagerly flagging the following UPDATE plan on the provider
itself:

    skipCredentialsValidation : nil -> false
    skipMetadataApiCheck      : nil -> true
    skipRegionValidation      : nil -> true

The new values are consistent with the defaults on the latest version so this plan seems to be OK.
The asserts are relaxed to permit update, but not replace plans on the provider when checking for
safe upgrades.
