+++ provider/provider_nodejs_test.go
@@ -66,21 +66,3 @@ func TestRegress3094(t *testing.T) {
 		providertest.WithSkippedUpgradeTestMode(providertest.UpgradeTestMode_PreviewOnly, "Not testing upgrades"),
 	)
 }
-
-func nodeTest(t *testing.T, dir string, opts ...providertest.Option) {
-	envRegion := getEnvRegion(t)
-	opts = append(opts,
-		providertest.WithConfig("aws:region", "INVALID_REGION"),
-		providertest.WithConfig("aws:envRegion", envRegion),
-	)
-	test(t, dir, opts...)
-}
-
-// This version of nodeTest does not aws:region INVALID_REGION manipulation.
-func simpleNodeTest(t *testing.T, dir string, opts ...providertest.Option) {
-	envRegion := getEnvRegion(t)
-	opts = append(opts,
-		providertest.WithConfig("aws:region", envRegion),
-	)
-	test(t, dir, opts...)
-}
