+++ provider/provider_test.go
@@ -4,8 +4,10 @@ package provider
 import (
 	"context"
 	"os"
+	"strings"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/pulumi/providertest"
@@ -38,7 +40,25 @@ func skipIfShort(t *testing.T) {
 	}
 }
 
+func withNoChangesInResourcesAndNoReplacementsInProvider() providertest.Option {
+	return providertest.WithDiffValidation(func(t *testing.T, diffs providertest.Diffs) {
+		for _, d := range diffs {
+			// For the explicit provider, prohibit replacements.
+			isProvider := strings.HasSuffix(string(d.URN), "aws::provider") ||
+				strings.Contains(string(d.URN), "pulumi:providers:aws")
+			if isProvider {
+				assert.Emptyf(t, d.Replaces, "Unexpected replacement plan for %#v", d)
+			} else { // For normal resources, prohibit any changes.
+				assert.Falsef(t, d.HasChanges, "Expected no change for %#v", d)
+			}
+		}
+	})
+}
+
 func test(t *testing.T, dir string, opts ...providertest.Option) {
+	if len(opts) == 0 {
+		opts = []providertest.Option{withNoChangesInResourcesAndNoReplacementsInProvider()}
+	}
 	skipIfShort(t)
 	opts = append(opts,
 		providertest.WithProviderName("aws"),
@@ -47,6 +67,30 @@ func test(t *testing.T, dir string, opts ...providertest.Option) {
 	)
 	ptest := providertest.NewProviderTest(dir, opts...)
 	ptest.Run(t)
+}
+
+func nodeTest(t *testing.T, dir string, opts ...providertest.Option) {
+	if len(opts) == 0 {
+		opts = []providertest.Option{withNoChangesInResourcesAndNoReplacementsInProvider()}
+	}
+	envRegion := getEnvRegion(t)
+	opts = append(opts,
+		providertest.WithConfig("aws:region", "INVALID_REGION"),
+		providertest.WithConfig("aws:envRegion", envRegion),
+	)
+	test(t, dir, opts...)
+}
+
+// This version of nodeTest does not aws:region INVALID_REGION manipulation.
+func simpleNodeTest(t *testing.T, dir string, opts ...providertest.Option) {
+	if len(opts) == 0 {
+		opts = []providertest.Option{withNoChangesInResourcesAndNoReplacementsInProvider()}
+	}
+	envRegion := getEnvRegion(t)
+	opts = append(opts,
+		providertest.WithConfig("aws:region", envRegion),
+	)
+	test(t, dir, opts...)
 }
 
 func TestUpgradeCoverage(t *testing.T) {
